VmbC API loader now automatically calls VmbStartup / VmbShutdown and reworked unit tests
The VmbCAPI class now automatically call VmbStartup, when it is loaded, and VmbShutdown, when it is destroyed. VmbCAPI behaves now like a singleton. When get_instance is called the first time the API is loaded and an instance of VmbCAPI is created. All later calls to get_instance will return the same instance independent of the argument libraryLoader. If all instance of VmbCAPI are destroyed, a call get_instance will create a new instance. The VmbCAPI class now also provides all functions of VmbC.
The unit test structure is now reworked and simplified. Also, a complete mock for the VmbC API is added, which can be later used for testing.