Reworked VimbaXCamera open by id mechanism
When a camera id is passed to the open function of the VimbaXCamera class, it now checks first if a matching camera can be found in the camera list. If no camera could be found, it falls back to just calling VmbCameraOpen with the given camera id string. The behaviour eliminates that an error message is shown, if the serial number is used for opening the camera.