Added streaming implementation and reworked logging
The vimbax_camera_node is now capable of streaming. The streaming is started automatically, when a subscriber is connected to the image topic (image_raw). It is also stopped again, when no node is subscribed. Added helpers to shift 10 and 12 bit formats left to be correctly aligned for the ROS 2 16 bit formats.
It is now possible to get the logger of the current node via the helper::get_logger method. By this all submodules log message are transmitted via the rosout topic.