Switch to multi thread executor
Previously simply rclcpp::spin was used to run the node. This is using a single threaded executor in the background. The single thread executor can only handle one service at the time and  if one service takes a bit longer to execute all other service are blocked. Therefore we now switched to a multi threaded executor which can handle multiple service request in parallel.