Changed the return type in the subscribe event service from int32 to Error, reworked event api and examples
The subscribe_event function of c++ and python event api now return a future of the subscription handle instead of returning the subscription handle directly. With this changes now it is possible to pass errors (e.g. when the event is not supported by the camera) to the user. This was not possible with the old implementation. The examples are also adjusted to the new api and also handle the subscribe errors.
Additionally the c++ event viewer examples is now correctly installed and available in the debian package.