Fixed adjust packet size is not called automatically for GigE cameras
With the has_feature method it is now possible to check if a feature is present in a specific GenTL module. This extensions is necessary, because the auto adjust packet size feature is a stream not remote device feature and previously has_feature was only working for remote device features. In case of the auto adjust packet size it always returned false and therefore the packet size was never adjusted correctly. To support this functionality the camera class stores now the feature information for each GenTL module in a map.